+++ projects/frontend/CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - No game launch if core doesn't match
 - Game not listed in lobby if password protected
+- Restore scrapped folder display
 
 ## [18.06.27] - 2018-06-27
 ### Added